#step2 [refactor] 패키징 분리
파일이 많아짐에 따라 model, view 패키징 분리
