+++ src/main/java/lotto/model/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 import java.util.List;
 import java.util.Random;
