+++ src/main/java/lotto/view/LottoInputView.java
@@ -1,4 +1,6 @@
-package lotto;
+package lotto.view;
+
+import lotto.model.Lotto;
 
 import java.util.InputMismatchException;
 import java.util.List;
