+++ src/main/java/lotto/view/LottoOutputView.java
@@ -1,4 +1,6 @@
-package lotto;
+package lotto.view;
+
+import lotto.model.Lotto;
 
 import java.util.List;
