+++ src/test/java/lotto/model/LottoGeneratorTest.java
@@ -1,7 +1,9 @@
-package lotto;
+package lotto.model;
 
+import lotto.model.Lotto;
+import lotto.model.LottoGenerator;
+import lotto.model.MockNumberGenerator;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
