+++ src/test/java/lotto/model/LottoTest.java
@@ -1,5 +1,8 @@
-package lotto;
+package lotto.model;
 
+import lotto.model.Lotto;
+import lotto.model.MockNumberGenerator;
+import lotto.model.NumberGenerator;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
